Fixed delete function, made some functions async
its been a while since working on this but previously delete didnt work. fixed delete to properly delete customer data after using the login button. also made all current functions that actually interact with the API to be async. before it would take two presses of the same button to get the new call to show in the console, the first click would always show data from the previous call